+++ lang/en/local_mailtest.php
@@ -32,11 +32,13 @@ $string['credit'] = 'Michael Milette - <a href="http://www.tngconsulting.ca/">TN
 $string['from'] = '{$a->type}: <strong>{$a->email}</strong> (<a href="{$a->url}">{$a->label}</a>)';
 $string['heading'] = 'Email configuration test';
 $string['recipientisrequired'] = 'You must specify the recipient\'s email address.';
-$string['errorsend'] = 'The test email message could not be delivered to the mail server.</p><p><strong>Recommendation:</strong></p><p>Check your Moodle <a href="{$a}" target="blank">Email settings</a>.';
-$string['errorcommunications'] = 'Moodle could not communicate with your mail server.</p><p><strong>Recommendation:</strong></p><p>Start by checking your Moodle <a href="{$a}" target="_blank">SMTP mail settings</a>.</p><p>If they look correct, check your SMTP Server and/or firewall settings to ensure that they are configured to accept SMTP connections from your Moodle web server and from your no-reply email address.';
+$string['errorsend'] = 'The test email message could not be delivered to the mail server.</p><p><strong>Recommendation:</strong></p><p>Check your Moodle <a href="{$a}" target="blank">Email settings</a>. For more help, see the FAQ section in the documentation.';
+$string['errorcommunications'] = 'Moodle could not communicate with your mail server.</p><p><strong>Recommendation:</strong></p><p>Start by checking your Moodle <a href="{$a}" target="_blank">SMTP mail settings</a>.</p><p>If they look correct, check your SMTP Server and/or firewall settings to ensure that they are configured to accept SMTP connections from your Moodle web server and from your no-reply email address. For more help, see the FAQ section in the documentation.';
 $string['fromemail'] = 'From email address';
 $string['toemail'] = 'To email address';
 $string['youremail'] = 'Your email address';
+$string['primaryadminemail'] = 'Primary admin email';
+$string['alwaysshowlog'] = 'Always show log of communications with mail server, even if there are no errors.';
 $string['sendtest'] = 'Send a test message';
 $string['sendmethod'] = 'Email send method';
 $string['sentmail'] = 'Moodle successfully delivered the test message to the SMTP mail server.';